reorganization of tf code
